Action Kotlin refactoring fixed
