+++ leader/src/main/scala/org/apache/amaterasu/leader/dsl/JobParser.scala
@@ -23,8 +23,8 @@ import com.fasterxml.jackson.databind.{JsonNode, ObjectMapper}
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory
 import org.apache.amaterasu.common.dataobjects.ActionData
 import org.apache.amaterasu.leader.execution.JobManager
-import org.apache.amaterasu.leader.execution.actions.{ErrorAction, SequentialAction}
-import org.apache.amaterasu.leader.execution.actions.Action
+import org.apache.amaterasu.leader.common.actions.{ErrorAction, SequentialAction}
+import org.apache.amaterasu.leader.common.execution.actions.Action
 import org.apache.curator.framework.CuratorFramework
 
 import scala.collection.JavaConverters._
